+++ R/knitr_utils.R
@@ -104,11 +104,57 @@ knitr_opts_current <- function(x, default = FALSE){
 #' The function is a utility to facilitate the retrieval of table
 #' options supported by the 'flextable', 'officedown' and of
 #' course 'officer' packages.
+#'
+#' These options should be set with `knitr::opts_chunk$set()`.
+#' The names and expected values are listed in the following sections.
+#'
+#' @section knitr chunk options for table captions:
+#'
+#' | **label**                                        |    **name**     | **value**  |
+#' |:-------------------------------------------------|:---------------:|:----------:|
+#' | caption id/bookmark                              | tab.id          |    NULL    |
+#' | caption                                          | tab.cap         |    NULL    |
+#' | display table caption on top of the table or not | tab.topcaption  |    TRUE    |
+#' | caption table sequence identifier.               | tab.lp          |   "tab:"   |
+#'
+#' @section knitr chunk options for Word table captions:
+#'
+#' | **label**                                               |    **name**     | **value**                 |
+#' |:--------------------------------------------------------|:---------------:|:-------------------------:|
+#' | Word stylename to use for table captions.               | tab.cap.style   |            NULL           |
+#' | prefix for numbering chunk (default to   "Table ").     | tab.cap.pre     |           Table           |
+#' | suffix for numbering chunk (default to   ": ").         | tab.cap.sep     |            " :"           |
+#' | title number depth                                      | tab.cap.tnd     |              0            |
+#' | separator to use between title number and table number. | tab.cap.tns     |             "-"           |
+#' | caption prefix formatting properties                    | tab.cap.fp_text | fp_text_lite(bold = TRUE) |
+#'
+#' @section knitr chunk options for Word tables:
+#'
+#' | **label**                                                     |    **name**  | **value**  |
+#' |:--------------------------------------------------------------|:------------:|:----------:|
+#' | the Word stylename to use for tables                          | tab.style    |    NULL    |
+#' | autofit' or 'fixed' algorithm.                                | tab.layout   |  "autofit" |
+#' | value of the preferred width of the table in percent (base 1).| tab.width    |      1     |
+#'
+#' @section knitr chunk options for data.frame with officedown:
+#'
+#' | **label**                                                     |    **name**     | **value**  |
+#' |:--------------------------------------------------------------|:---------------:|:----------:|
+#' | apply or remove formatting from the first row in the table    | first_row       |    TRUE    |
+#' | apply or remove formatting from the first column in the table | first_column    |    FALSE   |
+#' | apply or remove formatting from the last row in the table     | last_row        |    FALSE   |
+#' | apply or remove formatting from the last column in the table  | last_column     |    FALSE   |
+#' | don't display odd and even rows                               | no_hband        |    TRUE    |
+#' | don't display odd and even columns                            | no_vband        |    TRUE    |
+#'
 #' @return a list with following elements:
 #'
 #' * cap.style (default: NULL)
 #' * cap.pre (default: "Table ")
 #' * cap.sep (default: ":")
+#' * cap.tnd (default: 0)
+#' * cap.tns (default: "-")
+#' * cap.fp_text (default: `fp_text_lite(bold = TRUE)`)
 #' * id (default: NULL)
 #' * cap (default: NULL)
 #' * topcaption (default: TRUE)
@@ -127,7 +173,12 @@ knitr_opts_current <- function(x, default = FALSE){
 opts_current_table <- function() {
   tab.cap.style <- knitr_opts_current("tab.cap.style", default = NULL)
   tab.cap.pre <- knitr_opts_current("tab.cap.pre", default = "Table ")
-  tab.cap.sep <- knitr_opts_current("tab.cap.sep", default = ":")
+  tab.cap.sep <- knitr_opts_current("tab.cap.sep", default = ": ")
+
+  tab.cap.tnd <- knitr_opts_current("tab.cap.tnd", default = 0)
+  tab.cap.tns <- knitr_opts_current("tab.cap.tns", default = "-")
+  tab.cap.fp_text <- knitr_opts_current("tab.cap.fp_text", default = fp_text_lite(bold = TRUE))
+
   tab.cap <- knitr_opts_current("tab.cap", default = NULL)
   tab.topcaption <- knitr_opts_current("tab.topcaption", default = TRUE)
   tab.id <- knitr_opts_current("tab.id", default = NULL)
@@ -147,6 +198,9 @@ opts_current_table <- function() {
     cap.style = tab.cap.style,
     cap.pre = tab.cap.pre,
     cap.sep = tab.cap.sep,
+    cap.tnd = tab.cap.tnd,
+    cap.tns = tab.cap.tns,
+    cap.fp_text = tab.cap.fp_text,
     id = tab.id,
     topcaption = tab.topcaption,
     cap = tab.cap,